+++ CHANGELOG.md
@@ -29,6 +29,8 @@
   Print directories bold.
 * [#130](https://github.com/kowainik/summoner/issues/130):
   Show git revision version under `--version` command.
+* [#63](https://github.com/kowainik/summoner/issues/63):
+  Add `show --ghc` and `show --license` commands.
 
 1.0.6
 =====